server/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.4,<5
dj-database-url>=2.2,<3
psycopg[binary]>=3.2,<4
python-dotenv>=1.0,<2
requests>=2.32,<3
